move over core api.go and tests to the start of core/http
