+++ core/http/apt_suite_test.go
@@ -1,4 +1,4 @@
-package api_test
+package http_test
 
 import (
 	"testing"
